backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.39
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==42.0.4
pillow==11.1.0
python-multipart==0.0.9
starlette==0.36.3
bcrypt
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.6
python-jose==3.4.0
PyJWT==2.10.1
cachetools==5.5.2
<<<<<<< Updated upstream
# pip install google-api-python-client
=======
requests==2.32.3
>>>>>>> Stashed changes
